Fixing shapeless expressions
