+++ dataset/src/main/scala/frameless/TypedDataset.scala
@@ -2,6 +2,8 @@ package frameless
 
 import java.util
 
+import frameless.functions.CatalystExplodableCollection
+
 //import frameless.functions.CatalystExplodableCollection
 import frameless.ops._
 import org.apache.spark.rdd.RDD
@@ -1163,15 +1165,22 @@ class TypedDataset[T] protected[frameless](val dataset: Dataset[T])(implicit val
     }
   }
 
-  def explode[A, TRep <: HList, V[_], OutRep <: HList, Out]
+  /**
+    * Explodes (flattens) a single column at a time. It only compiles if the type of column supports this operation.
+    *
+    * @param column the column we wish to explode/flatten
+    */
+  def explode[A, TRep <: HList, V[_], OutMod <: HList, OutModValues <: HList, Out]
   (column: Witness.Lt[Symbol])
   (implicit
    i0: TypedColumn.Exists[T, column.T, V[A]],
-   i5: TypedEncoder[A],
-   i1: LabelledGeneric.Aux[T, TRep],
-   i2: Modifier.Aux[TRep, column.T, V[A], A, OutRep],
-   i3: Tupler.Aux[OutRep, Out],
-   i4: TypedEncoder[Out]
+   i1: TypedEncoder[A],
+   i2: CatalystExplodableCollection[V],
+   i3: LabelledGeneric.Aux[T, TRep],
+   i4: Modifier.Aux[TRep, column.T, V[A], A, OutMod],
+   i5: Values.Aux[OutMod, OutModValues],
+   i6: Tupler.Aux[OutModValues, Out],
+   i7: TypedEncoder[Out]
   ): TypedDataset[Out] = {
     val df = dataset.toDF()
     import org.apache.spark.sql.functions.{explode => sparkExplode}
