+++ dataset/src/main/scala/frameless/functions/UnaryFunctions.scala
@@ -42,6 +42,8 @@ trait UnaryFunctions {
     *
     * apache/spark
     */
+  @deprecated("Use explode() from the TypedDataset instead. This method will result in " +
+    "runtime error if applied to two columns in the same select statement.", "0.6.2")
   def explode[T, A: TypedEncoder, V[_] : CatalystExplodableCollection](column: TypedColumn[T, V[A]]): TypedColumn[T, A] =
     new TypedColumn[T, A](sparkFunctions.explode(column.untyped))
 }
@@ -71,6 +73,7 @@ object CatalystExplodableCollection {
   implicit def explodableVector: CatalystExplodableCollection[Vector] = new CatalystExplodableCollection[Vector] {}
   implicit def explodableArray: CatalystExplodableCollection[Array] = new CatalystExplodableCollection[Array] {}
   implicit def explodableList: CatalystExplodableCollection[List] = new CatalystExplodableCollection[List] {}
+  implicit def explodableSeq: CatalystExplodableCollection[Seq] = new CatalystExplodableCollection[Seq] {}
 }
 
 trait CatalystSortableCollection[V[_]] {
